+++ packages/module-source/index.js
@@ -1,1 +1,4 @@
+// eslint-disable-next-line import/export
+export * from './src/external.types.js';
+
 export { ModuleSource } from './src/module-source.js';