+++ packages/random/uint.js
@@ -1,0 +1,7 @@
+export {
+  randomUint8,
+  randomUint16,
+  randomUint24,
+  randomUint32,
+  randomUint53,
+} from './src/uint.js';